+++ src/main/resources/templates/emp/changePassword.html
@@ -46,11 +46,13 @@
             <div class="container">
                 <h1>비밀번호 변경</h1>
                 <form id="passwordForm" th:action="@{/emp/password}" method="post">
-                    <input type="password" class="form-control passwordInput" id="currentPassword" name="currentPassword"
+                    <input type="password" class="form-control passwordInput" id="currentPassword"
+                           name="currentPassword"
                            placeholder="현재 비밀번호" required>
                     <input type="password" class="form-control passwordInput" id="newPassword" name="newPassword"
                            placeholder="새 비밀번호" required>
-                    <input type="password" class="form-control passwordInput" id="newPasswordVerify" name="newPasswordVerify"
+                    <input type="password" class="form-control passwordInput" id="newPasswordVerify"
+                           name="newPasswordVerify"
                            placeholder="새 비밀번호 확인" required>
                     <!-- 버튼 그룹 -->
                     <div class="button-group">
@@ -61,6 +63,7 @@
             </div>
         </main>
         <script>
+            const contextPath = /*[[@{/}]]*/ '';
             $(document).ready(function () {
                 $('#passwordForm').submit(function (e) {
                     e.preventDefault();
@@ -90,7 +93,7 @@
                     }
 
                     $.ajax({
-                        url: '/emp/password',
+                        url: contextPath + '/api/emp/password',
                         method: 'POST',
                         data: {
                             currentPassword: currentPassword,
