+++ src/main/resources/templates/emp/empInfo.html
@@ -135,7 +135,7 @@
                         <tr>
                             <th rowspan="4" id="photo">
                         <span id="showImage">
-                            <img th:src="@{/emp/profile-image(empId=${empInfoDTO.emp.empId})}"
+                            <img th:src="@{/api/emp/profile-image(empId=${empInfoDTO.emp.empId})}"
                                  width="200px" alt="profile" id="over_view"/>
                                 <th:block th:if="${empInfoDTO.emp.getImgUUID()!=null}">
                                 <img th:src="@{/image/cancel.png}" alt="파일삭제" width="20px" class="remove_img"
