+++ src/main/resources/templates/layout.html
@@ -94,7 +94,7 @@
             </li>
             <li class="nav-item dropdown">
                 <a class="nav-link dropdown-toggle" id="navbarDropdown" href="#" role="button" data-bs-toggle="dropdown"
-                   aria-expanded="false"><img th:src="@{/emp/profile-image(empId=${#authentication.name})}"
+                   aria-expanded="false"><img th:src="@{/api/emp/profile-image(empId=${#authentication.name})}"
                                               width="30px"></a>
                 <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
                     <li><a class="dropdown-item" th:href="@{/emp/password}">비밀번호 변경</a></li>
