+++ src/main/resources/templates/post/boardView.html
@@ -165,7 +165,7 @@
                         </tr>
                         <tr>
                             <td colspan="3" style="padding-top: 10px;">
-                                <img th:src="@{|/emp/profile-image?empId=${id}|}" alt="프로필 이미지" class="icon viewIcon">
+                                <img th:src="@{|/api/emp/profile-image?empId=${id}|}" alt="프로필 이미지" class="icon viewIcon">
                                 <div class="username">[[${empMap.get(postDetailDTO.postWriter)}]]</div>
                                 <div class="date"> [[${postDetailDTO.getFormattedPostDate()}]]</div>
                             </td>
@@ -244,7 +244,7 @@
                         </ul>
                         <div class="comment-write-container">
                             <span class="comment-photo">
-                                <img th:src="@{|/emp/profile-image?empId=${id}|}" class="profileImg" width=36 height=36>
+                                <img th:src="@{|/api/emp/profile-image?empId=${id}|}" class="profileImg" width=36 height=36>
                             </span>
                             <div class="comment-write">
                                 <div class="comment-write-area">
